Move schemata utilities to common folder
